ML Codes/requirements.txt
boto3==1.36.16
botocore==1.36.16
certifi==2024.12.14
charset-normalizer==3.4.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
jmespath==1.0.1
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
smmap==5.0.2
tzdata==2025.1
urllib3==2.3.0
